pyproject.toml
[project]
name = "model-context-protocol-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=2.14.2",
    "google-genai>=1.57.0",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
